Attempt to update up_meica_fixes.
Widespread changes, and I suspect the error we're attempting to
fix is being introduced by upstream in more areas of the script.

However, I don't really understand what we're attempting to fix, and
don't have the time right now to figure it out. This is a quick best effort,
and I'll come back to verify/fix as necessary.