Update CMake patch for build changes
Did not port the "building" of afni_papers.h from afni_papers.txt.

Upstream already ships (and updates) afni_papers.h. Plus, quotize is
built, and then used (assumed to be in path) to convert the txt into a
header file...

Not worth providing a build-dep during build for something that is
already done for us.